refactor flashmessage for eslint
