+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "github-username-regex": "^1.0.0",
+    "prop-types": "^15.6.2",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-scripts": "1.1.4"
